#89 cleanup
