+++ calendar/src/main/kotlin/calendar/logic.kt
@@ -89,95 +89,3 @@ fun generateMonth(_time: LocalDate): MutableList<Week> {
 	} while(time.month == _time.month && time.dayOfMonth > 1)
 	return weeks
 }
-
-abstract class DBObservable<T>: DBObservableD<T, T>() {
-	override fun convertFrom(value: T): T = value
-	
-	override fun convertTo(value: T): T = value
-	
-}
-
-abstract class DBDateObservable: DBObservableD<LocalDate, Long>() {
-	override fun convertFrom(value: LocalDate): Long = value.toUTCEpochMinute()
-	override fun convertTo(value: Long): LocalDate = Timing.fromUTCEpochMinuteToLocalDateTime(value).toLocalDate()
-}
-
-abstract class DBDateTimeObservable: DBObservableD<LocalDateTime, Long>() {
-	override fun convertFrom(value: LocalDateTime): Long = value.toUTCEpochMinute()
-	override fun convertTo(value: Long): LocalDateTime = Timing.fromUTCEpochMinuteToLocalDateTime(value)
-}
-
-abstract class DBObservableD<T, DB>: ObjectPropertyBase<T>() {
-	private var loaded = false
-	override fun getBean(): Any = ""//TODO("Not yet implemented")
-	
-	override fun getName(): String = ""//TODO("Not yet implemented")
-	
-	private fun reload() {
-		transaction {
-			set(convertTo(abstractGet()))
-		}
-	}
-	
-	abstract fun abstractGet(): DB
-	
-	abstract fun abstractSet(dat: DB)
-	
-	override fun get(): T {
-		if(!loaded)
-			reload()
-		return super.get()
-	}
-	
-	override fun set(newValue: T) {
-		transaction {
-			abstractSet(convertFrom(newValue))
-		}
-		super.set(newValue)
-	}
-	
-	fun set(v: Property<T>) {
-		set(v.value)
-	}
-	
-	abstract fun convertFrom(value: T): DB
-	
-	abstract fun convertTo(value: DB): T
-	
-	override fun hashCode(): Int = value.hashCode()
-	
-	/**
-	 * used for equals methods of DB Classes
-	 */
-	override fun equals(other: Any?): Boolean {
-		if(this === other)
-			return true
-		if(other !is DBObservableD<*, *>)
-			return false
-		if(value != other.value)
-			return false
-		
-		return true
-	}
-	
-	fun clone(): Property<T> = SimpleObjectProperty(value)
-	
-	override fun toString(): String = "[DBO(${get()})]"
-}
-
-/**
- * only get() is allowed to get its value, this list must always be bound to
- *
- * set and setValue are also blocked, only remove and add
- */
-open class DBObservableList<T: Entity<*>>(private val table: EntityClass<*, T>): ObservableListWrapper<T>(mutableListOf()) {
-	private var loaded = false
-	
-	fun reload() {
-		return transaction {
-			super.setAll(table.all().toList())
-		}
-	}
-	
-	override fun toString(): String = "[DBObservableList value: ${super.toString()}]"
-}
